Move transform block back into question/content
