+++ tests/functional/spec/custom-page-titles.spec.js
@@ -1,5 +1,5 @@
 import HubPage from "../base_pages/hub.page.js";
-import DateOfBirthPage from "../generated_pages/custom_page_titles/date-of-birth.page.js"
+import DateOfBirthPage from "../generated_pages/custom_page_titles/date-of-birth.page.js";
 import IndividualInterstitialPage from "../generated_pages/custom_page_titles/individual-interstitial.page.js";
 import ListCollectorAddPage from "../generated_pages/custom_page_titles/list-collector-add.page.js";
 import ListCollectorEditPage from "../generated_pages/custom_page_titles/list-collector-edit.page.js";
